requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-core
python-dotenv
requests
streamlit-mic-recorder
duckduckgo-search
ddgs
